+++ packages/deploy-aws/bin/_pipeline-options.js
@@ -12,4 +12,6 @@ module.exports = {
   repository: process.env.GITHUB_REPO,
   ssmParameter: process.env.GITHUB_SSM_PARAMETER,
   computeTypeSize: process.env.COMPUTE_TYPE_SIZE || 'SMALL',
+  buildWorkers: parseInt(process.env.BUILD_WORKER_COUNT || "1", 10),
+  excludedPages: process.env.PAGES_TO_EXCLUDE
 }