+++ packages/deploy-aws/src/lib/get_stage_configs.ts
@@ -87,8 +87,13 @@ function getInstall(params: StageConfigParams) {
 }
 
 function getBuild(params: StageConfigParams) {
+  // Fixes an issue where the max file watch count is exceeded, triggering ENOSPC
+  // https://stackoverflow.com/questions/22475849/node-js-error-enospc#32600959
+  const increaseFileWatcherLimit = addStateChangeToCmd(
+    `echo fs.inotify.max_user_watches=524288 | sudo tee -a /etc/sysctl.conf && sudo sysctl -p`
+  )
   const runBuild = addStateChangeToCmd('npm run build')
-  return getCommandConfig(params, [runBuild])
+  return getCommandConfig(params, [increaseFileWatcherLimit, runBuild])
 }
 
 function getTest(params: StageConfigParams) {
@@ -108,8 +113,12 @@ function getTestEnd2End(params: StageConfigParams) {
 }
 
 function getDeploy(params: StageConfigParams) {
+  const mergeCommands = [...Array(params.buildWorkers).keys()].map(i =>
+    addStateChangeToCmd(`cp -Rf  $CODEBUILD_SRC_DIR_dist${i}/dist/* dist`)
+  )
   const commands = [
-    addStateChangeToCmd('cp -Rf $CODEBUILD_SRC_DIR_dist/dist dist'),
+    addStateChangeToCmd('mkdir dist'),
+    ...mergeCommands,
     checkState(),
     addStateChangeToCmd('npm run deploy'),
     checkState(),
@@ -156,6 +165,7 @@ function getEnv(params: StageConfigParams) {
     STAGE: params.stage,
     TARGET_URL: params.pipelineLink,
     TEST_PIPELINE_CONTEXT: 'continuous-integration/seagull',
+    ...(params.excludedPages ? { PAGES_TO_EXCLUDE: params.excludedPages } : {}),
   }
   return { 'parameter-store': parameterStore, variables }
 }