+++ packages/deploy-aws/src/templates/seagull_pipeline.ts
@@ -8,7 +8,9 @@ import { StageConfigParams } from '../types'
 interface SeagullPipelineProps {
   appPath: string
   branch: string
+  buildWorkers: 1 | 2 | 3 | 4
   computeTypeSize: 'SMALL' | 'MEDIUM' | 'LARGE'
+  excludedPages?: string
   githubToken?: string
   poll: boolean
   profile: string
@@ -36,6 +38,8 @@ export class SeagullPipeline {
   stage: string
   githubToken?: string
   actions: string[]
+  buildWorkers: 1 | 2 | 3 | 4
+  excludedPages?: string
 
   constructor(props: SeagullPipelineProps) {
     this.appPath = props.appPath
@@ -49,8 +53,10 @@ export class SeagullPipeline {
     this.repository = props.repository
     this.ssmParam = props.ssmParameter
     this.githubToken = props.githubToken
+    this.buildWorkers = props.buildWorkers
     const propsSSMHandler = props.handlers && props.handlers.ssmHandler
     this.ssm = propsSSMHandler || new SSMHandler()
+    this.excludedPages = props.excludedPages
     this.actions = [
       'cloudformation:*',
       'cloudfront:*',
@@ -100,7 +106,9 @@ export class SeagullPipeline {
     const pipelineLink = `${pipelineDomain}${pipelinePath}`
     const stageConfigParams: StageConfigParams = {
       branch: gitData.branch,
+      buildWorkers: this.buildWorkers,
       computeTypeSize: this.computeTypeSize,
+      excludedPages: this.excludedPages,
       owner: gitData.owner,
       pipeline,
       pipelineLink,