+++ packages/libraries/src/index.ts
@@ -3,5 +3,6 @@ export * from './Library'
 export * from './fileSystemHelper'
 export * from './loggingHelper'
 export * from './basicError'
+export * from './objectHelper'
 import * as Typings from './typings'
 export { Typings }