+++ packages/seed/src/error.ts
@@ -1,13 +1,13 @@
 import { BasicError } from '@seagull/libraries'
-import { FixtureStorage } from './fixtureStorage'
+import { FixtureStorage } from './fixture-storage'
 
 export class SeedError extends BasicError {
   /**
    * Creates a new seed error
    * @param message human readable error message
-   * @param seed seed fixture storage
+   * @param details fixture storage or inner error object
    */
-  constructor(message: string, public seed: FixtureStorage<any> | any) {
+  constructor(message: string, public details?: FixtureStorage<any> | Error) {
     super('SeedError', message)
   }
 }