+++ packages/services-http/src/mode/base.ts
@@ -1,3 +1,4 @@
+import { SeedableService } from '@seagull/seed'
 import { injectable } from 'inversify'
 import { RequestInit, Response } from 'node-fetch'
 import 'reflect-metadata'
@@ -7,7 +8,7 @@ import { IHttp, RequestInitBase, RequestInitGet } from '../interface'
  * Http (default) base implementation.
  */
 @injectable()
-export abstract class HttpBase implements IHttp {
+export abstract class HttpBase extends SeedableService implements IHttp {
   abstract fetch(url: string, init?: RequestInit | undefined): Promise<Response>
 
   async get(url: string, init?: RequestInitGet): Promise<Response> {