+++ packages/soap/src/index.ts
@@ -1,2 +1,0 @@
-export * from './soap'
-export * from './soap-di-module'