+++ packages/soap/src/soap-di-module.ts
@@ -1,6 +1,0 @@
-import { ContainerModule } from 'inversify'
-import { SoapClientSupplier } from './soap'
-
-export const soapDIModule = new ContainerModule(bind => {
-  bind<SoapClientSupplier>(SoapClientSupplier).toSelf()
-})