+++ packages/testing/src/service_test.ts
@@ -1,10 +1,17 @@
+// tslint:disable: member-ordering
 import { SetMode } from '@seagull/mode'
+import { TestScope } from '@seagull/seed'
 import { Container } from 'inversify'
+import { IBeforeAndAfterContext, IHookCallbackContext } from 'mocha'
+import { context } from 'mocha-typescript'
 
 export abstract class ServiceTest {
+  @context private mocha!: IBeforeAndAfterContext & IHookCallbackContext
+
   abstract serviceModules: any[]
   abstract services: any[]
   injector = new Container()
+  stateful = false
 
   /** Implement your logic to run before each single test here!
    * if you override before(), you compromise internal logic of InjectableTest */
@@ -18,6 +25,13 @@ export abstract class ServiceTest {
   before() {
     new SetMode('environment', 'pure').execute()
     this.injector = new Container()
+    if (this.stateful) {
+      const suite = this.mocha.currentTest!.parent!.title
+      const test = this.mocha.currentTest!.title
+      this.injector
+        .bind(TestScope)
+        .toDynamicValue(() => new TestScope(suite, test))
+    }
     for (const diMod of this.serviceModules) {
       this.injector.load(diMod)
     }